+++ scenex/ek/scenex/components/transform_2d.cpp
@@ -5,21 +5,33 @@
 namespace ek {
 
 void begin_transform(const transform_2d& transform) {
-    if (transform.user_matrix != nullptr) {
-        transform.matrix = *(transform.user_matrix);
-    } else {
-        transform.matrix.set(transform.scale, transform.skew);
-    }
-    draw2d::state
-            .save_transform()
+    const_cast<transform_2d&>(transform).updateLocalMatrix();
+    draw2d::state.save_transform()
             .combine_color(transform.color_multiplier, transform.color_offset)
-            .translate(transform.origin)
-            .concat_matrix(transform.matrix)
-            .translate(-transform.origin);
+            .concat_matrix(transform.matrix);
 }
 
 void end_transform() {
     draw2d::state.restore_transform();
+}
+
+void transform_2d::updateLocalMatrix() {
+    const auto x = position.x + origin.x;
+    const auto y = position.y + origin.y;
+    const auto xx = -origin.x - pivot.x;
+    const auto yy = -origin.y - pivot.y;
+
+    const auto ra = cosf(skew.y) * scale.x;
+    const auto rb = sinf(skew.y) * scale.x;
+    const auto rc = -sinf(skew.x) * scale.y;
+    const auto rd = cosf(skew.x) * scale.y;
+
+    matrix.a = ra;
+    matrix.b = rb;
+    matrix.c = rc;
+    matrix.d = rd;
+    matrix.tx = x + ra * xx + rc * yy;
+    matrix.ty = y + rd * yy + rb * xx;
 }
 
 }