+++ scenex/ek/scenex/systems/layout_system.hpp
@@ -26,32 +26,29 @@ public:
 
     layout_wrapper& hard(float x, float y) {
         auto& transform = ecs::get_or_create<transform_2d>(e_);
-        horizontal(x, transform.matrix.tx - (designCanvasRect.x + designCanvasRect.width * x));
-        vertical(y, transform.matrix.ty - (designCanvasRect.y + designCanvasRect.height * y));
+        horizontal(x, transform.position.x - (designCanvasRect.x + designCanvasRect.width * x));
+        vertical(y, transform.position.y - (designCanvasRect.y + designCanvasRect.height * y));
         return *this;
     }
 
     layout_wrapper& hard_y(float y = 0.0f) {
         auto& transform = ecs::get_or_create<transform_2d>(e_);
-        vertical(y, transform.matrix.ty - (designCanvasRect.y + designCanvasRect.height * y));
+        vertical(y, transform.position.y - (designCanvasRect.y + designCanvasRect.height * y));
         return *this;
     }
 
     layout_wrapper& horizontal(float multiplier = 0.0f, float offset = 0.0f) {
-        l_.align_x = true;
-        l_.x = {multiplier, offset};
+        l_.enableAlignX(multiplier, offset);
         return *this;
     }
 
     layout_wrapper& vertical(float multiplier = 0.0f, float offset = 0.0f) {
-        l_.align_y = true;
-        l_.y = {multiplier, offset};
+        l_.enableAlignY(multiplier, offset);
         return *this;
     }
 
     layout_wrapper& fill(bool horizontal = true, bool vertical = true) {
-        l_.fill_x = horizontal;
-        l_.fill_y = vertical;
+        l_.fill(horizontal, vertical);
         return *this;
     }
 
@@ -59,6 +56,7 @@ public:
         l_.fill_extra = rc;
         return *this;
     }
+
 private:
     ecs::entity e_;
     layout_t& l_;